requirements.txt
matplotlib 
numpy 
Pillow 
torch 
torchvision 
